+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.0.14] - 2022-12-15
+
+### Added
+
+- Show or live tail docker container logs
+
+
 ## [1.0.13] - 2022-11-18
 
 ### Added